Make Radius required, use Black as colour in Circle
Both Polygon and Polyline use Black as the default colour, making them visible on the Map when used without specifying a colour.
Circle by default draws a 0-radius circle. If the `radius` is specified, it still draws a `Transparent` circle on the map, making it not visible.
Update the `Circle` function to *require* the radius to be specified, and change the default colour to `Color.Black`, as used in other Composable for Maps.